requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.4
Click==8.0
itsdangerous==2.1.0
Jinja2==3.0.0
MarkupSafe==2.1.1
Werkzeug==2.2.2
flask-login
Flask-WTF
email-validator
flask-migrate